backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncmy==0.2.9
alembic==1.14.1
pydantic-settings==2.7.1
pyjwt==2.10.1
bcrypt==4.2.1
python-dotenv==1.0.1
python-multipart==0.0.18
structlog==24.4.0
